fix code duplication
