+++ app/src/main/java/com/android/mySwissDorm/ui/listing/EditListingScreen.kt
@@ -1,5 +1,6 @@
 package com.android.mySwissDorm.ui.listing
 
+import EditListingViewModel
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
