+++ app/src/main/java/com/android/mySwissDorm/ui/listing/EditListingViewModel.kt
@@ -1,151 +1,40 @@
-package com.android.mySwissDorm.ui.listing
-
 import android.util.Log
 import androidx.lifecycle.viewModelScope
-import com.android.mySwissDorm.model.map.Location
 import com.android.mySwissDorm.model.map.LocationRepository
 import com.android.mySwissDorm.model.map.LocationRepositoryProvider
 import com.android.mySwissDorm.model.rental.RentalListing
 import com.android.mySwissDorm.model.rental.RentalListingRepository
 import com.android.mySwissDorm.model.rental.RentalListingRepositoryProvider
 import com.android.mySwissDorm.model.rental.RentalStatus
-import com.android.mySwissDorm.model.rental.RoomType
 import com.android.mySwissDorm.model.residency.ResidenciesRepository
 import com.android.mySwissDorm.model.residency.ResidenciesRepositoryProvider
-import com.android.mySwissDorm.model.residency.Residency
-import com.android.mySwissDorm.ui.InputSanitizers
-import com.android.mySwissDorm.ui.InputSanitizers.FieldType
-import com.android.mySwissDorm.ui.utils.BaseLocationSearchViewModel
+import com.android.mySwissDorm.ui.listing.BaseListingFormViewModel
 import com.google.firebase.Firebase
 import com.google.firebase.Timestamp
 import com.google.firebase.auth.auth
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
-/**
- * Immutable UI model for the Edit-Listing screen.
- *
- * This state is the single source of truth for the screen fields and is validated via the
- * centralized [InputSanitizers]. It is produced and updated by [EditListingViewModel] and consumed
- * by the composables.
- *
- * @property title Listing title shown in the header and cards.
- * @property residencyName Selected [Residency] to which the listing belongs.
- * @property residencies Available list of residencies for the dropdown.
- * @property price Monthly price as a user-typed string (normalized while typing).
- * @property housingType Selected [RoomType] (e.g., STUDIO, ROOM, etc.).
- * @property sizeSqm Area as a user-typed string in square meters (normalized while typing).
- * @property startDate Desired availability start date for the listing.
- * @property description Free-form description entered by the user.
- * @property pickedImages List of image URLs chosen for the listing.
- * @property mapLat Optional latitude derived from the residency or map picker.
- * @property mapLng Optional longitude derived from the residency or map picker.
- * @property errorMsg Optional human-readable error to display in the UI. Documentation generated by
- *   AI
- */
-data class EditListingUIState(
-    val title: String = "",
-    val residencyName: String = "",
-    val residencies: List<Residency>,
-    val price: String = "",
-    val housingType: RoomType,
-    val sizeSqm: String = "",
-    val startDate: Timestamp,
-    val description: String = "",
-    val pickedImages: List<String> = emptyList(),
-    val mapLat: Double? = null,
-    val mapLng: Double? = null,
-    val errorMsg: String? = null,
-    val customLocationQuery: String = "",
-    val customLocation: Location? = null,
-    val locationSuggestions: List<Location> = emptyList(),
-    val showCustomLocationDialog: Boolean = false,
-) {
-  /**
-   * Derived validity flag for the whole form.
-   *
-   * Uses centralized validators to ensure that each text field is syntactically valid and within
-   * allowed ranges. This keeps UI logic "dumb" and uniform.
-   */
-  val isFormValid: Boolean
-    get() {
-      val titleOk = InputSanitizers.validateFinal<String>(FieldType.Title, title).isValid
-      val sizeOk = InputSanitizers.validateFinal<Double>(FieldType.RoomSize, sizeSqm).isValid
-      val priceOk = InputSanitizers.validateFinal<Int>(FieldType.Price, price).isValid
-      val descOk = InputSanitizers.validateFinal<String>(FieldType.Description, description).isValid
-      return titleOk && descOk && sizeOk && priceOk
-    }
-}
-
-/**
- * ViewModel backing the Edit-Listing screen.
- *
- * Responsibilities:
- * - Load the existing listing by id and project it into [EditListingUIState].
- * - Keep the UI state up to date as the user edits fields (normalized while typing).
- * - Validate and persist edits through [RentalListingRepository].
- * - Provide a simple error channel via [EditListingUIState.errorMsg].
- *
- * Dependencies default to providers but can be injected for testing.
- */
 class EditListingViewModel(
-    private val rentalListingRepository: RentalListingRepository =
-        RentalListingRepositoryProvider.repository,
-    private val residenciesRepository: ResidenciesRepository =
-        ResidenciesRepositoryProvider.repository,
-    override val locationRepository: LocationRepository = LocationRepositoryProvider.repository
-) : BaseLocationSearchViewModel() {
+    rentalListingRepository: RentalListingRepository = RentalListingRepositoryProvider.repository,
+    residenciesRepository: ResidenciesRepository = ResidenciesRepositoryProvider.repository,
+    locationRepository: LocationRepository = LocationRepositoryProvider.repository
+) :
+    BaseListingFormViewModel(
+        rentalListingRepository = rentalListingRepository,
+        residenciesRepository = residenciesRepository,
+        locationRepository = locationRepository) {
+
   override val logTag = "EditListingViewModel"
 
-  /**
-   * Backing state for the screen. Initialized with a placeholder [Residency] and defaults for all
-   * form fields.
-   */
-  private val _uiState =
-      MutableStateFlow(
-          EditListingUIState(
-              residencies = listOf(), startDate = Timestamp.now(), housingType = RoomType.STUDIO))
-
-  /** Public, read-only view of the current UI state. */
-  val uiState: StateFlow<EditListingUIState> = _uiState.asStateFlow()
-
-  /** Clears the current error message, if any. */
-  fun clearErrorMsg() {
-    _uiState.value = _uiState.value.copy(errorMsg = null)
-  }
-
-  /** Sets a new error message to be displayed by the UI. */
-  private fun setErrorMsg(errorMsg: String) {
-    _uiState.value = _uiState.value.copy(errorMsg = errorMsg)
-  }
-
-  /**
-   * Loads a listing by its ID and updates the UI state.
-   *
-   * @param rentalPostID The ID of the listing to be loaded.
-   */
   fun getRentalListing(rentalPostID: String) {
     viewModelScope.launch {
       try {
         val listing = rentalListingRepository.getRentalListing(rentalPostID)
-        val currentResidencies = _uiState.value.residencies
         val isPrivateAccommodation = listing.residencyName == "Private Accommodation"
-        if (isPrivateAccommodation) {
-          null
-        } else {
-          try {
-            residenciesRepository.getResidency(listing.residencyName)
-          } catch (e: Exception) {
-            null
-          }
-        }
-        // Get location from the listing (now stored in the model)
         val listingLocation = listing.location
 
         _uiState.value =
-            EditListingUIState(
+            uiState.value.copy(
                 title = listing.title,
                 residencyName = listing.residencyName,
                 price = listing.pricePerMonth.toString(),
@@ -156,60 +45,18 @@ class EditListingViewModel(
                 pickedImages = listing.imageUrls,
                 mapLat = listingLocation.latitude,
                 mapLng = listingLocation.longitude,
-                customLocation =
-                    if (isPrivateAccommodation) {
-                      // For Private Accommodation, use the location from the listing
-                      listingLocation
-                    } else {
-                      null
-                    },
-                customLocationQuery =
-                    if (isPrivateAccommodation) {
-                      listingLocation.name
-                    } else {
-                      ""
-                    },
-                errorMsg = null,
-                residencies = currentResidencies)
-        // If residencies haven't been loaded yet, load them now
-        if (currentResidencies.isEmpty()) {
-          loadResidencies()
-        }
+                customLocation = if (isPrivateAccommodation) listingLocation else null,
+                customLocationQuery = if (isPrivateAccommodation) listingLocation.name else "",
+                errorMsg = null)
       } catch (e: Exception) {
-        Log.e("EditListingViewModel", "Error loading listing by ID: $rentalPostID", e)
+        Log.e(logTag, "Error loading listing by ID: $rentalPostID", e)
         setErrorMsg("Failed to load listing: ${e.message}")
       }
     }
   }
 
-  /**
-   * Persists edited values to the repository.
-   *
-   * Note: This is an internal helper that performs the actual repository call. Callers should
-   * generally use [editRentalListing] which validates form state first.
-   *
-   * @param id The ID of the listing to be edited.
-   * @param listing The [RentalListing] object containing the new values.
-   */
-  private fun editRentalListingToRepository(id: String, listing: RentalListing) {
-    viewModelScope.launch {
-      try {
-        rentalListingRepository.editRentalListing(rentalPostId = id, newValue = listing)
-      } catch (e: Exception) {
-        Log.e("EditListingViewModel", "Error editing listing", e)
-        setErrorMsg("Failed to edit rental listing: ${e.message}")
-      }
-    }
-  }
-
-  /**
-   * Validates the current UI state and, if valid, writes the updated listing.
-   *
-   * @param id The listing identifier.
-   * @return `true` if the update was queued (form valid), `false` otherwise.
-   */
   fun editRentalListing(id: String): Boolean {
-    val state = _uiState.value
+    val state = uiState.value
     if (!state.isFormValid) {
       setErrorMsg("At least one field is not valid")
       return false
@@ -219,214 +66,44 @@ class EditListingViewModel(
         Firebase.auth.currentUser?.uid
             ?: throw IllegalStateException("User must be authenticated to edit a listing")
 
-    // Determine location: use custom location for Private Accommodation, otherwise use residency
-    // location
-    val location =
-        if (state.residencyName == "Private Accommodation") {
-          if (state.customLocation != null) {
-            state.customLocation
-          } else if (state.mapLat != null && state.mapLng != null) {
-            // Fallback: create location from mapLat/mapLng if customLocation is null
-            Location(
-                name = state.customLocationQuery.ifEmpty { "Custom Location" },
-                latitude = state.mapLat,
-                longitude = state.mapLng)
-          } else {
-            setErrorMsg("Please select a location for Private Accommodation")
-            return false
-          }
-        } else {
-          // For regular residencies, get location from the residency
-          val residency = state.residencies.find { it.name == state.residencyName }
-          if (residency != null) {
-            residency.location
-          } else if (state.mapLat != null && state.mapLng != null) {
-            // Fallback: use mapLat/mapLng if residency not found
-            Location(name = state.residencyName, latitude = state.mapLat, longitude = state.mapLng)
-          } else {
-            setErrorMsg("Could not determine location for the selected residency")
-            return false
-          }
-        }
+    val location = resolveLocationOrNull() ?: return false
 
-    editRentalListingToRepository(
-        id = id,
-        listing =
-            RentalListing(
-                uid = id,
-                ownerId = uid,
-                postedAt = Timestamp.now(),
-                residencyName = state.residencyName,
-                title = state.title,
-                roomType = state.housingType,
-                pricePerMonth = state.price.toDouble(),
-                areaInM2 = state.sizeSqm.toDouble().toInt(),
-                startDate = state.startDate,
-                description = state.description,
-                imageUrls = state.pickedImages,
-                status = RentalStatus.POSTED,
-                location = location))
+    val listing =
+        RentalListing(
+            uid = id,
+            ownerId = uid,
+            postedAt = Timestamp.now(),
+            residencyName = state.residencyName,
+            title = state.title,
+            roomType = state.housingType,
+            pricePerMonth = state.price.toDouble(),
+            areaInM2 = state.sizeSqm.toDouble().toInt(),
+            startDate = state.startDate,
+            description = state.description,
+            imageUrls = state.pickedImages,
+            status = RentalStatus.POSTED,
+            location = location)
+
+    viewModelScope.launch {
+      try {
+        rentalListingRepository.editRentalListing(rentalPostId = id, newValue = listing)
+      } catch (e: Exception) {
+        Log.e(logTag, "Error editing listing", e)
+        setErrorMsg("Failed to edit rental listing: ${e.message}")
+      }
+    }
     clearErrorMsg()
     return true
   }
 
-  init {
-    loadResidencies()
-  }
-  /**
-   * Deletes a listing document by its ID.
-   *
-   * @param rentalPostID The ID of the listing to be deleted.
-   */
   fun deleteRentalListing(rentalPostID: String) {
     viewModelScope.launch {
       try {
         rentalListingRepository.deleteRentalListing(rentalPostId = rentalPostID)
       } catch (e: Exception) {
-        Log.e("EditListingViewModel", "Error deleting listing", e)
+        Log.e(logTag, "Error deleting listing", e)
         setErrorMsg("Failed to delete listing: ${e.message}")
       }
     }
-  }
-
-  /**
-   * Loads available residencies to populate the residency dropdown.
-   *
-   * Errors are surfaced through [EditListingUIState.errorMsg].
-   */
-  private fun loadResidencies() {
-    viewModelScope.launch {
-      try {
-        val residencies = residenciesRepository.getAllResidencies()
-        _uiState.value = _uiState.value.copy(residencies = residencies)
-      } catch (e: Exception) {
-        _uiState.value = _uiState.value.copy(errorMsg = e.message)
-      }
-    }
-  }
-
-  // --- Update handlers: normalize while typing via central rules ---
-
-  /**
-   * Updates [EditListingUIState.title] using normalized input.
-   *
-   * @param title Raw user input for the title.
-   */
-  fun setTitle(title: String) {
-    val norm = InputSanitizers.normalizeWhileTyping(FieldType.Title, title)
-    _uiState.value = _uiState.value.copy(title = norm)
-  }
-
-  /**
-   * Updates the selected [Residency].
-   *
-   * @param String The new residency.
-   */
-  fun setResidency(residencyName: String) {
-    val currentState = _uiState.value
-    val isPrivateAccommodation = residencyName == "Private Accommodation"
-
-    if (isPrivateAccommodation) {
-      // For Private Accommodation, keep custom location if already set, otherwise clear it
-      _uiState.value =
-          currentState.copy(
-              residencyName = residencyName,
-              // Keep custom location if it exists, otherwise it stays null
-          )
-    } else {
-      // For regular residencies, set location from the residency
-      val residency = currentState.residencies.find { it.name == residencyName }
-      _uiState.value =
-          currentState.copy(
-              residencyName = residencyName,
-              customLocation = null, // Clear custom location for regular residencies
-              customLocationQuery = "",
-              mapLat = residency?.location?.latitude,
-              mapLng = residency?.location?.longitude)
-    }
-  }
-
-  fun getCityName(residencyName: String): String {
-    return _uiState.value.residencies.find { it.name == residencyName }?.city ?: "Unknown"
-  }
-
-  /**
-   * Updates [EditListingUIState.price] using normalized input.
-   *
-   * @param price Raw user input for monthly price.
-   */
-  fun setPrice(price: String) {
-    val norm = InputSanitizers.normalizeWhileTyping(FieldType.Price, price)
-    _uiState.value = _uiState.value.copy(price = norm)
-  }
-
-  /**
-   * Updates the selected [RoomType].
-   *
-   * @param housingType The new room type.
-   */
-  fun setHousingType(housingType: RoomType) {
-    _uiState.value = _uiState.value.copy(housingType = housingType)
-  }
-
-  /**
-   * Updates [EditListingUIState.sizeSqm] using normalized input.
-   *
-   * @param sizeSqm Raw user input for area in square meters.
-   */
-  fun setSizeSqm(sizeSqm: String) {
-    val norm = InputSanitizers.normalizeWhileTyping(FieldType.RoomSize, sizeSqm)
-    _uiState.value = _uiState.value.copy(sizeSqm = norm)
-  }
-
-  /**
-   * Updates the listing availability start date.
-   *
-   * @param startDate New [Timestamp] value.
-   */
-  fun setStartDate(startDate: Timestamp) {
-    _uiState.value = _uiState.value.copy(startDate = startDate)
-  }
-
-  /**
-   * Updates [EditListingUIState.description] using normalized input.
-   *
-   * @param description Raw user input for the description.
-   */
-  fun setDescription(description: String) {
-    val norm = InputSanitizers.normalizeWhileTyping(FieldType.Description, description)
-    _uiState.value = _uiState.value.copy(description = norm)
-  }
-
-  // --- Custom location methods (BaseLocationSearchViewModel) ---
-  override fun updateStateWithQuery(query: String) {
-    _uiState.value = _uiState.value.copy(customLocationQuery = query)
-  }
-
-  override fun updateStateWithSuggestions(suggestions: List<Location>) {
-    _uiState.value = _uiState.value.copy(locationSuggestions = suggestions)
-  }
-
-  override fun updateStateWithLocation(location: Location) {
-    _uiState.value =
-        _uiState.value.copy(
-            customLocation = location,
-            customLocationQuery = location.name,
-            mapLat = location.latitude,
-            mapLng = location.longitude)
-  }
-
-  public override fun updateStateShowDialog(currentLocation: Location?) {
-    _uiState.value =
-        _uiState.value.copy(
-            showCustomLocationDialog = true,
-            customLocation = currentLocation,
-            customLocationQuery = currentLocation?.name ?: "")
-  }
-
-  public override fun updateStateDismissDialog() {
-    _uiState.value =
-        _uiState.value.copy(
-            showCustomLocationDialog = false, customLocationQuery = "", customLocation = null)
   }
 }
